+++ packages/terra-action-footer/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 2.56.0 - (December 8, 2020)
+
+* Changed
+  * Minor dependency version bump
+
 ## 2.55.0 - (November 17, 2020)
 
 * Changed