+++ packages/terra-badge/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 3.46.0 - (December 8, 2020)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.45.0 - (November 17, 2020)
 
 * Changed